Revert "Add all the files of the project"
This reverts commit eef98f268c686b08393158fc044aa5317695f61a.

That was a done before I delete the remote GitHub repo and copied it to
local folder. After the new repo I copied the local folder files into
the new local repo. Accidentally copied the another projects as well
which should not be done as I want the project Form-Controls cleaned
without files and commits from Wireframe.